requirements.txt
graphics
pillow
pygame
numpy